Put components & main.jsx in react dir
Also rename main.jsx --> index.jsx
