+++ app/react/components/AdminIssueItem.jsx
