+++ app/react/components/AdvocateIssueItem.jsx
