+++ app/react/components/AvailableDashboard.jsx
