+++ app/react/components/AvailableIssueItem.jsx
