+++ app/react/components/Footer.jsx
