+++ app/react/index.jsx
@@ -3,11 +3,12 @@ import React from 'react'
 import { Router, Route, IndexRedirect, browserHistory } from 'react-router'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
-import { receiveIssues, receiveAdvocateIssues, receiveOpenIssues } from './redux/reducers/issues'
-import { receiveVolunteers } from './redux/reducers/volunteers'
-import { receiveAdvocates } from './redux/reducers/advocates'
 
-import store from './redux/store'
+import store from '../redux/store'
+import { receiveIssues, receiveAdvocateIssues, receiveOpenIssues } from '../redux/reducers/issues'
+import { receiveVolunteers } from '../redux/reducers/volunteers'
+import { receiveAdvocates } from '../redux/reducers/advocates'
+
 import App from './components/App'
 import Landing from './components/Landing'
 import Hotline from './components/Hotline'
